refactor: moving slides to providers
