+++ apps/ubuntu_bootstrap/test/installer_test.dart
@@ -9,8 +9,8 @@ import 'package:ubuntu_bootstrap/app.dart';
 import 'package:ubuntu_bootstrap/l10n.dart';
 import 'package:ubuntu_bootstrap/pages.dart';
 import 'package:ubuntu_bootstrap/pages/autoinstall/autoinstall_model.dart';
+import 'package:ubuntu_bootstrap/providers/slides_provider.dart';
 import 'package:ubuntu_bootstrap/services.dart';
-import 'package:ubuntu_bootstrap/slides/slides_provider.dart';
 import 'package:ubuntu_provision/ubuntu_provision.dart';
 import 'package:ubuntu_wizard/ubuntu_wizard.dart';
 import 'package:yaru_test/yaru_test.dart';
